// src/app/core/enums/keyboard.enums.ts
export enum KeyboardKey {
  ARROW_UP = 'ArrowUp',
  ARROW_DOWN = 'ArrowDown',
  ARROW_LEFT = 'ArrowLeft',
  ARROW_RIGHT = 'ArrowRight',
  TAB = 'Tab',
  ENTER = 'Enter',
  ESCAPE = 'Escape',
}

export enum KeyboardAction {
  MOVE_UP = 'move-up',
  MOVE_DOWN = 'move-down',
  MOVE_LEFT = 'move-left',
  MOVE_RIGHT = 'move-right',
  MOVE_NEXT = 'move-next',
  ADD_SUBCATEGORY = 'add-subcategory',
  NONE = 'none',
}